+++ lib/flutter_waya.dart
@@ -6,7 +6,9 @@ export 'package:flutter_easyrefresh/easy_refresh.dart';
 export 'package:flutter_localizations/flutter_localizations.dart';
 export 'package:http_parser/http_parser.dart';
 
+export 'components/components.dart';
 export 'constant/constant.dart';
 export 'extension/extension.dart';
+export 'spinKit/spinKit.dart';
 export 'utils/utils.dart';
 export 'widgets/widgets.dart';