+++ lib/widgets/widgets.dart
@@ -1,5 +1,6 @@
-export '../components/components.dart';
-export '../spinKit/spinKit.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
 export 'icon.dart';
 export 'input_text.dart';
 export 'list_wheel.dart';
@@ -12,3 +13,25 @@ export 'scroll/sliver/sliver.dart';
 export 'tab_bar.dart';
 export 'text.dart';
 export 'universal.dart';
+
+class SystemUiOverlayStyleLight extends SystemUiOverlayStyle {
+  const SystemUiOverlayStyleLight()
+      : super(
+            systemNavigationBarColor: Colors.transparent,
+            systemNavigationBarDividerColor: Colors.transparent,
+            statusBarColor: Colors.transparent,
+            systemNavigationBarIconBrightness: Brightness.light,
+            statusBarIconBrightness: Brightness.light,
+            statusBarBrightness: Brightness.dark);
+}
+
+class SystemUiOverlayStyleDark extends SystemUiOverlayStyle {
+  const SystemUiOverlayStyleDark()
+      : super(
+            systemNavigationBarColor: Colors.transparent,
+            systemNavigationBarDividerColor: Colors.transparent,
+            statusBarColor: Colors.transparent,
+            systemNavigationBarIconBrightness: Brightness.dark,
+            statusBarIconBrightness: Brightness.dark,
+            statusBarBrightness: Brightness.light);
+}